+++ src/space.cpp
@@ -1168,7 +1168,7 @@ void space_text_balance_nested_parens(void)
 }
 
 
-size_t space_needed(chunk_t *first, chunk_t *second)
+uint32_t space_needed(chunk_t* first, chunk_t* second)
 {
    LOG_FUNC_ENTRY();
    LOG_FMT(LSPACE, "%s\n", __func__);
@@ -1177,7 +1177,7 @@ size_t space_needed(chunk_t *first, chunk_t *second)
    switch (do_space(first, second, min_sp))
    {
       case AV_ADD:    /* fallthrough */
-      case AV_FORCE:  return((size_t)(max(1, min_sp)));
+      case AV_FORCE:  return((uint32_t)(max(1, min_sp)));
       case AV_REMOVE: return(0);
       case AV_IGNORE: /* fallthrough */
       default:        return(second->orig_col > (first->orig_col + first->len()));
@@ -1185,7 +1185,7 @@ size_t space_needed(chunk_t *first, chunk_t *second)
 }
 
 
-size_t space_col_align(chunk_t *first, chunk_t *second)
+uint32_t space_col_align(chunk_t* first, chunk_t* second)
 {
    LOG_FUNC_ENTRY();
    assert(are_valid(first, second));
@@ -1202,7 +1202,7 @@ size_t space_col_align(chunk_t *first, chunk_t *second)
    argval_t av = do_space(first, second, min_sp);
 
    LOG_FMT(LSPACE, "%s: av=%d, ", __func__, av);
-   size_t coldiff;
+   uint32_t coldiff;
    if (first->nl_count > 0)
    {
       LOG_FMT(LSPACE, "nl_count=%zu, orig_col_end=%u", first->nl_count, first->orig_col_end);
